Grid code onderverdeeld in een aparte .cpp bestand met een header file.
